Prerender, use vercel adapter, disable PUT action for now
