+++ sites/svelte.dev/.gitignore
@@ -9,3 +9,4 @@
 /src/routes/_components/Supporters/contributors.js
 /src/routes/_components/Supporters/donors.jpg
 /src/routes/_components/Supporters/donors.js
+.vercel
