+++ sites/svelte.dev/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@supabase/supabase-js": "^2.11.0",
+    "@sveltejs/adapter-vercel": "^2.4.1",
     "@sveltejs/repl": "^0.2.0",
     "cookie": "^0.5.0",
     "devalue": "^4.3.0",
@@ -28,7 +29,6 @@
   },
   "devDependencies": {
     "@resvg/resvg-js": "^2.4.1",
-    "@sveltejs/adapter-auto": "^2.0.0",
     "@sveltejs/kit": "^1.12.0",
     "@sveltejs/site-kit": "^3.3.6",
     "@sveltejs/vite-plugin-svelte": "^2.0.3",
