+++ sites/svelte.dev/src/routes/(authed)/repl/[id].json/+server.js
@@ -1,4 +1,4 @@
-import { dev } from '$app/environment';
+import { building, dev } from '$app/environment';
 import { client } from '$lib/db/client';
 import * as gist from '$lib/db/gist';
 import * as session from '$lib/db/session';
@@ -6,16 +6,14 @@ import { get_example } from '$lib/server/examples';
 import { get_examples_data, get_examples_list } from '$lib/server/examples/get-examples';
 import { error, json } from '@sveltejs/kit';
 
+export const prerender = 'auto';
+
 const UUID_REGEX = /^[0-9a-f]{8}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{4}-?[0-9a-f]{12}$/;
 
 const examples_data = get_examples_data();
 
-const examples = new Set(
-	get_examples_list(examples_data)
-		.map((category) => category.examples)
-		.flat()
-		.map((example) => example.slug)
-);
+/** @type {Set<string>} */
+let examples;
 
 /** @param {import('$lib/server/examples/types').ExamplesData[number]['examples'][number]['files'][number][]} files  */
 function munge(files) {
@@ -40,16 +38,28 @@ function munge(files) {
 }
 
 export async function GET({ params }) {
-	if (examples.has(params.id)) {
-		const example = get_example(examples_data, params.id);
+	// Currently, these pages(that are in examples/) are prerendered. To avoid making any FS requests,
+	// We prerender examples pages during build time. That means, when something like `/repl/hello-world.json`
+	// is accessed, this function won't be run at all, as it will be served from the filesystem
+	if (building || dev) {
+		examples = new Set(
+			get_examples_list(examples_data)
+				.map((category) => category.examples)
+				.flat()
+				.map((example) => example.slug)
+		);
 
-		return json({
-			id: params.id,
-			name: example.title,
-			owner: null,
-			relaxed: false, // TODO is this right? EDIT: It was example.relaxed before, which no example return to my knowledge. By @PuruVJ
-			components: munge(example.files),
-		});
+		if (examples.has(params.id)) {
+			const example = get_example(examples_data, params.id);
+
+			return json({
+				id: params.id,
+				name: example.title,
+				owner: null,
+				relaxed: false, // TODO is this right? EDIT: It was example.relaxed before, which no example return to my knowledge. By @PuruVJ
+				components: munge(example.files),
+			});
+		}
 	}
 
 	if (dev && !client) {
@@ -78,13 +88,13 @@ export async function GET({ params }) {
 	});
 }
 
-// TODO reimplement as an action
-export async function PUT({ params, request }) {
-	const user = await session.from_cookie(request.headers.get('cookie'));
-	if (!user) throw error(401, 'Unauthorized');
+// // TODO reimplement as an action
+// export async function PUT({ params, request }) {
+// 	const user = await session.from_cookie(request.headers.get('cookie'));
+// 	if (!user) throw error(401, 'Unauthorized');
 
-	const body = await request.json();
-	await gist.update(user, params.id, body);
+// 	const body = await request.json();
+// 	await gist.update(user, params.id, body);
 
-	return new Response(undefined, { status: 204 });
-}
+// 	return new Response(undefined, { status: 204 });
+// }
